+++ openhands/server/auth.py
