+++ openhands/server/middleware.py
@@ -11,9 +11,9 @@ from fastapi.responses import JSONResponse
 from starlette.middleware.base import BaseHTTPMiddleware
 from starlette.types import ASGIApp
 
-from openhands.server.shared import session_manager, config
 from openhands.core.logger import openhands_logger as logger
-
+from openhands.server.shared import config, session_manager
+from openhands.server.auth import get_sid_from_token, sign_token
 
 
 class LocalhostCORSMiddleware(CORSMiddleware):
